src/requirements.txt
tqdm>=4.67.1
pandas>=2.2.2
numpy>=2.0.0,<3.0.0
torch>=2.7.0
torchvision>=0.22.0
scikit-learn>=1.6.1
scipy>=1.15.2
Pillow>=11.1.0
matplotlib>=3.10.0
transformers>=4.50.3
kornia>=0.8.0
timm>=1.0.15
einops>=0.8.1
safetensors>=0.7.0
shapely>=2.1.1
opencv-python>=4.11.0.86
